Fix 'Invalid count argument' errors in user module
Using `var.tag_id` in the count argument of `pagerduty_tag_assignment`
results in an 'Invalid count argument' error because terraform can't
calculate the number of resources to create before the tag resources
have been created.

The fix is to add an `add_tag` argument to the user module and calculate
whether to set it to 0 or 1 based on the value of var.tag_name, which is
known before any resources are created.